requirements.txt
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.1.0
customtkinter==5.2.0
numpy==1.24.3
qrcode==7.4.2

